+++ reckon-core/src/main/groovy/org/ajoberstar/reckon/core/Scope.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ajoberstar.semver.vcs;
+package org.ajoberstar.reckon.core;
 
 import com.github.zafarkhaja.semver.Version;
