+++ reckon-core/src/test/groovy/org/ajoberstar/reckon/core/GrgitVcsSpec.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ajoberstar.semver.vcs.grgit
+package org.ajoberstar.reckon.core
 
 import java.nio.file.Files
 import java.security.SecureRandom
