+++ reckon-gradle/src/main/resources/META-INF/gradle-plugins/org.ajoberstar.semver-vcs-base.properties
@@ -1,1 +1,1 @@
-implementation-class=org.ajoberstar.semver.vcs.gradle.SemverVcsPlugin
+implementation-class=org.ajoberstar.reckon.gradle.GrgitVcsPlugin
