--- reckon-gradle/src/main/resources/META-INF/gradle-plugins/org.ajoberstar.semver-vcs-grgit.properties
@@ -1,1 +1,0 @@
-implementation-class=org.ajoberstar.semver.vcs.gradle.grgit.GrgitVcsPlugin
